requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-cors-headers==4.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-decouple==3.8
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
